streamlit-apps/msa_to_txt/requirements.txt
biopython
streamlit
